Move InferredMemberName simplification helpers to shared layer
