+++ src/Workspaces/VisualBasic/Portable/Simplification/Reducers/VisualBasicInferredMemberNameReducer.Rewriter.vb
@@ -18,6 +18,15 @@ Namespace Microsoft.CodeAnalysis.VisualBasic.Simplification
             End Sub
 
             Private ReadOnly s_simplifyTupleName As Func(Of SimpleArgumentSyntax, SemanticModel, OptionSet, CancellationToken, SimpleArgumentSyntax) = AddressOf SimplifyTupleName
+            Private ReadOnly s_simplifyNamedFieldInitializer As Func(Of NamedFieldInitializerSyntax, SemanticModel, OptionSet, CancellationToken, SyntaxNode) = AddressOf SimplifyNamedFieldInitializer
+
+            Private Function SimplifyNamedFieldInitializer(node As NamedFieldInitializerSyntax, arg2 As SemanticModel, optionSet As OptionSet, arg4 As CancellationToken) As SyntaxNode
+                If CanSimplifyNamedFieldInitializer(node) Then
+                    Return SyntaxFactory.InferredFieldInitializer(node.Expression).WithTriviaFrom(node)
+                End If
+
+                Return node
+            End Function
 
             Private Function SimplifyTupleName(
                 node As SimpleArgumentSyntax,
